+++ src/Encamina.Enmarcha.Data.EntityFramework/AsyncReadRepositoryBase.cs
@@ -42,6 +42,6 @@ public abstract class AsyncReadRepositoryBase<TEntity> : IAsyncReadRepository<TE
     }
 
     /// <inheritdoc/>
-    public virtual async Task<TEntity> GetByIdAsync<TEntityId>(TEntityId id, CancellationToken cancellationToken)
+    public virtual async Task<TEntity?> GetByIdAsync<TEntityId>(TEntityId id, CancellationToken cancellationToken)
         => await dbContext.GetSet<TEntity>().FindAsync(new object[] { id }, cancellationToken);
 }