+++ CHANGELOG.md
@@ -82,6 +82,7 @@ Previous classification is not required if changes are simple or all belong to t
   - `Encamina.Enmarcha.AspNet`
   - `Encamina.Enmarcha.Bot`
   - `Encamina.Enmarcha.Core`
+  - `Encamina.Enmarcha.Data`
  
 ## [8.1.5]